+++ R/character.R
@@ -16,14 +16,18 @@
 #' substrEnd("12345", 1, 1)
 #' substrEnd("12345", 1, 2)
 #' substrEnd("12345", 2, 3)
-#' 
+#'
 substrEnd <- function(x, start, endDiff) {
+    assertthat::assert_that(is.character(x))
+    assertthat::assert_that(is.numeric(start))
+    assertthat::assert_that(is.numeric(endDiff))
     substr(x, start, nchar(x) - endDiff)
 }
 
 #' Enclose String with Specified Characters
 #'
-#' This function encloses a string with specified characters on the left and the right.
+#' This function encloses a string with specified characters on the left
+#' and the right.
 #'
 #' @param x A character string to enclose.
 #' @param left A character string to prepend.
@@ -35,7 +39,10 @@ substrEnd <- function(x, start, endDiff) {
 #' @examples
 #' enclose("text", "[", "]") # returns "[text]"
 enclose <- function(x, left, right) {
-    paste(left, x, right, sep = "")
+    assertthat::assert_that(is.character(x))
+    assertthat::assert_that(is.character(left) && length(left) == 1)
+    assertthat::assert_that(is.character(right) && length(right) == 1)
+    paste0(left, x, right)
 }
 
 #' Enclose String with Brackets
@@ -50,6 +57,7 @@ enclose <- function(x, left, right) {
 #' @examples
 #' encloseBr("text") # returns "(text)"
 encloseBr <- function(x) {
+    assertthat::assert_that(is.character(x))
     enclose(x, "(", ")")
 }
 
@@ -58,8 +66,11 @@ encloseBr <- function(x) {
 #' This function retrieves a character at a specific index from a string.
 #'
 #' @param x A character string.
-#' @param index The index of the character to retrieve.
-#' 
+#' @param index The index of the character to retrieve. If it is length 1,
+#' then the same character is retrieved for all elements of `x`. Otherwise,
+#' if it is the same length as `x`, then the character at each index is
+#' retrieved.
+#'
 #' @return The character at the specified index.
 #' @export
 #' @keywords character
@@ -67,5 +78,11 @@ encloseBr <- function(x) {
 #' # Get the character at index 2
 #' getChar("hello", 2)
 getChar <- function(x, index) {
+
+    assertthat::assert_that(is.character(x))
+    assertthat::assert_that(
+        is.numeric(index) && (length(index) == length(x) || length(index) == 1)
+    )
+
     substr(x, index, index)
 }