+++ R/dataInitialization.R
@@ -132,6 +132,7 @@ tableToNumeric <- function(x) {
 #' vec <- c(a = 1, b = 2, c = 3)
 #' namedNumericToTable(vec)
 namedNumericToTable <- function(x) {
+    assertthat::assert_that(is.numeric(x) && !is.null(names(x)))
     output <- as.integer(x)
     names(output) <- names(x)
     output <- as.table(output)
@@ -156,7 +157,11 @@ namedNumericToTable <- function(x) {
 #' # Create a hash table with keys and initial value of 0
 #' createHash(c("a", "b", "c"), 0)
 createHash <- function(keys, initVals = NULL) {
+
     if (missing(keys)) return(hash::hash())
+    assertthat::assert_that(is.character(keys))
+    assertthat::assert_that(is.null(initVals) || length(initVals) == 1)
+
     keys <- unique(keys)
     numkeys <- length(keys)
     switch(as.character(numkeys),